# src/modules/coreHttpApi/openapi.yml
openapi: 3.0.0
servers:
  - url: /
    description: Local Connector API
info:
  version: "0.4"
  title: Connector
  description: |
    This API is work in progress and might be changed without further notice on any time.

    You can retrieve the [yaml](/docs/yaml) or [json](/docs/json) representation of this specification by following the corresponding links.

    You can also view this documenation in [rapidoc](/docs/rapidoc).

  license:
    name: MIT
    url: https://raw.githubusercontent.com/nmshd/cns-connector/main/LICENSE
  contact:
    name: j&s-soft GmbH
    email: info@js-soft.com
    url: https://www.js-soft.com/
paths:
...
  /api/v1/Account/IdentityInfo:
    get:
      operationId: getIdentityInfo
      description: Shows the Account Information.
      tags:
        - Account
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    nullable: false
                    $ref: "#/components/schemas/IdentityInfo"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
  /api/v1/Account/SyncInfo:
    get:
      operationId: getSyncInfo
      description: show information about the last completed sync run
      tags:
        - Account
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      lastSyncRun:
                        type: object
                        properties:
                          completedAt:
                            type: string
                            format: date-time
                        required:
                          - syncTime
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
  /api/v1/Account/Sync:
    post:
      operationId: sync
      description: >
        Syncs the Connector messages and relationships with the Backbone.
        Checks for new relationships as well as incoming changes of existing ones.
        Checks for new or updated Messages.
        Returns all affected relationships and messages.
      tags:
        - Account
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      relationships:
                        type: array
                        items:
                          $ref: "#/components/schemas/Relationship"
                      messages:
                        type: array
                        items:
                          $ref: "#/components/schemas/Message"
                    required:
                      - messages
                      - relationships
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"

  # ------------------- Files -------------------
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    Address:
      type: string
      format: Address
      minLength: 35
      maxLength: 36
      example: id_________________________________
    DeviceID:
      type: string
      format: DeviceID
      minLength: 20
      maxLength: 20
      example: DVC_________________
    FileID:
      type: string
      format: FileID
      minLength: 20
      maxLength: 20
      example: FIL_________________
      description: The ID of a file.
    MessageID:
      type: string
      format: MessageID
      minLength: 20
      example: MSG_________________
      maxLength: 20
      description: The ID of a message.
    RelationshipID:
      type: string
      format: RelationshipID
      minLength: 20
      maxLength: 20
      example: REL_________________
      description: The ID of a relationship.
    RelationshipChangeID:
      type: string
      format: RelationshipChangeID
      minLength: 20
      maxLength: 20
      example: RCH_________________
      description: The ID of a relationship change.
    RelationshipTemplateID:
      type: string
      format: RelationshipTemplateID
      minLength: 20
      maxLength: 20
      example: RLT_________________
      description: The ID of a relationship template.
...
    IdentityInfo:
      type: object
      properties:
        address:
          allOf:
            - $ref: "#/components/schemas/Address"
          description: The enmeshed address of the Connector.
          nullable: false
        publicKey:
          type: string
          description: The public key of the Connector used to digitally sign data.
          nullable: false
      required:
        - address
        - publicKey
...
    Recipient:
      type: object
      description: Information about the recipient of a message.
      properties:
        address:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The address of this recipient.
      required:
        - address
    Message:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/MessageID"
          nullable: false
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The ID of the sender of this message.
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
          description: The ID of the device this message was sent with.
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this message was received by the platform.
        recipients:
          type: array
          description: Information about the recipients of this message.
          items:
            $ref: "#/components/schemas/Recipient"
        relationshipIds:
          type: array
          description: If you are the sender of this message, the array will contain the IDs of the `Relationship`s to all recipients. If you are a recipient, the array will only contain the ID of the `Relationship` to the sender.
          items:
            type: string
        content:
          allOf:
            - $ref: "#/components/schemas/MessageContent"
          nullable: false
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/FileID"
      required:
        - id
        - createdBy
        - createdByDevice
        - createdAt
        - recipients
        - relationshipIds
        - content
...
    MessageContent:
      type: object
    Relationship:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipID"
          nullable: false
          description: The ID of the relationship.
        template:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplate"
          nullable: false
        status:
          type: string
          enum: [Pending, Active, Rejected, Revoked, Terminated]
          nullable: false
          description: The status of the relationship
        peer:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The address of the peer identity.
        changes:
          type: array
          items:
            $ref: "#/components/schemas/RelationshipChange"
          nullable: false
        lastMessageSentAt:
          type: string
          format: date-time
          nullable: true
          description: A timestamp that describes when the last message was sent to the peer.
        lastMessageReceivedAt:
          type: string
          format: date-time
          nullable: true
          description: A timestamp that describes when the last message was received from the peer.
      required:
        - id
        - template
        - status
        - peer
        - changes
    RelationshipChange:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipChangeID"
          nullable: false
        type:
          type: string
          nullable: false
          enum: [Creation, Termination]
        status:
          type: string
          nullable: false
          enum: [Pending, Rejected, Revoked, Accepted]
        request:
          $ref: "#/components/schemas/RelationshipChangeRequest"
        response:
          $ref: "#/components/schemas/RelationshipChangeResponse"
      required:
        - id
        - type
        - status
        - request
    RelationshipChangeRequest:
      type: object
      properties:
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this `RelationshipChange` was created on the platform.
          readOnly: true
        content:
          $ref: "#/components/schemas/RelationshipChangeRequestContent"
      required:
        - createdBy
        - createdByDevice
        - createdAt
        - content
    RelationshipChangeRequestContent:
      type: object
    RelationshipChangeResponse:
      type: object
      properties:
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this `RelationshipChange` was created on the platform.
          readOnly: true
        content:
          $ref: "#/components/schemas/RelationshipChangeResponseContent"
      required:
        - createdBy
        - createdByDevice
        - createdAt
        - content
    RelationshipChangeResponseContent:
      type: object
    RelationshipTemplateContent:
      type: object
    RelationshipTemplate:
      type: object
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplateID"
          nullable: false
          readOnly: true
        maxNumberOfRelationships:
          type: integer
          format: int32
          nullable: true
          description: The maximum number of relationships that can be established with this template. Omitting this property leads to an infinite number of possible relationships.
        isOwn:
          type: boolean
          description: Describes whether the template was created by the connector (true) or whether it was loaded from another identity (false).
          nullable: false
        createdBy:
          allOf:
            - $ref: "#/components/schemas/Address"
          nullable: false
          description: The address of the identity who created this template.
        createdByDevice:
          allOf:
            - $ref: "#/components/schemas/DeviceID"
          nullable: false
          description: The address of the identity who created this template.
        createdAt:
          type: string
          format: date-time
          nullable: false
          description: A timestamp that describes when this relationship template was created on the platform.
        expiresAt:
          type: string
          format: date-time
          description: A timestamp that describes when this relationship template expires. Expired templates cannot be used to create relationship requests anymore.
        content:
          allOf:
            - $ref: "#/components/schemas/RelationshipTemplateContent"
          nullable: false
      required:
        - id
        - isOwn
        - createdBy
        - createdByDevice
        - createdAt
        - expiresAt
        - content
...
    ErrorContent:
      type: object
      properties:
        error:
          type: object
          properties:
            id:
              description: A unique ID for the error instance.
              type: string
            code:
              description: >
                The error code.  

                There are some generic error codes like: 
                  - error.connector.recordNotFound
                  - error.connector.unauthorized
                  - error.connector.forbidden
                  - error.connector.validation.missingProperty
                  - error.connector.validation.invalidPropertyValueLength
                  - error.connector.validation.invalidPropertyValue

                Further there are route specific error codes.
              type: string
            message:
              description: An error message having more details about what went wrong. This message is not localized.
              type: string
            docs:
              description: A link to additional documentation about the error.
              type: string
            time:
              description: A timestamp that describes when the error occured.
              type: string
          required:
            - id
            - code
            - message
            - docs
            - time
      required:
        - error
    HeaderContent_X-Response-Duration-ms:
      description: The time in milliseconds the request took to be processed by the platform.
      type: number
    HeaderContent_X-Response-Time:
      description: A timestamp that describes when the response was sent to the client.
      type: string
      format: number
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
...